+++ src/main/java/gift/controller/req/ProductRequest.java
@@ -1,4 +1,4 @@
-package gift.controller.res;
+package gift.controller.req;
 
 import gift.model.Product;
